fix : DB 연결 수정
