+++ src/main/java/chess/domain/dao/DBConnection.java
@@ -1,0 +1,33 @@
+package chess.domain.dao;
+
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+
+public class DBConnection {
+    private Connection con = null;
+    private String server = "localhost:13306";
+    private String database = "chess_game";
+    private String option = "?useSSL=false&serverTimezone=UTC";
+    private String userName = "root";
+    private String password = "root";
+
+    public Connection getConnection() {
+        try {
+            Class.forName("com.mysql.cj.jdbc.Driver");
+        } catch (ClassNotFoundException e) {
+            System.err.println(" !! JDBC Driver load 오류: " + e.getMessage());
+            e.printStackTrace();
+        }
+
+        try {
+            con = DriverManager.getConnection("jdbc:mysql://" + server + "/" + database + option, userName, password);
+            System.out.println("정상적으로 연결되었습니다.");
+        } catch (SQLException e) {
+            System.err.println("연결 오류:" + e.getMessage());
+            e.printStackTrace();
+        }
+
+        return con;
+    }
+}
