+++ src/main/java/chess/domain/dao/PieceDAO.java
@@ -1,6 +1,7 @@
 package chess.domain.dao;
 
 import chess.domain.board.Board;
+import chess.domain.dto.PiecesDto;
 import chess.domain.game.BlackTurn;
 import chess.domain.game.ChessGame;
 import chess.domain.game.State;
@@ -15,39 +16,10 @@ import java.util.List;
 
 public class PieceDAO {
 
-    public Connection getConnection() {
-        Connection con = null;
-        String server = "localhost:13306"; // MySQL 서버 주소
-        String database = "chess_game"; // MySQL DATABASE 이름
-        String option = "?useSSL=false&serverTimezone=UTC";
-        String userName = "root"; //  MySQL 서버 아이디
-        String password = "root"; // MySQL 서버 비밀번호
+    DBConnection dbConnection;
 
-        try {
-            Class.forName("com.mysql.cj.jdbc.Driver");
-        } catch (ClassNotFoundException e) {
-            System.err.println(" !! JDBC Driver load 오류: " + e.getMessage());
-            e.printStackTrace();
-        }
-
-        try {
-            con = DriverManager.getConnection("jdbc:mysql://" + server + "/" + database + option, userName, password);
-            System.out.println("정상적으로 연결되었습니다.");
-        } catch (SQLException e) {
-            System.err.println("연결 오류:" + e.getMessage());
-            e.printStackTrace();
-        }
-
-        return con;
-    }
-
-    public void closeConnection(Connection con) {
-        try {
-            if (con != null)
-                con.close();
-        } catch (SQLException e) {
-            System.err.println("con 오류:" + e.getMessage());
-        }
+    public PieceDAO() {
+        dbConnection = new DBConnection();
     }
 
     public void save(final Piece piece) {
@@ -56,10 +28,9 @@ public class PieceDAO {
         if ("WHITE".equals(color)) {
             shape = shape.toLowerCase();
         }
-        try {
-            Connection con = getConnection();
-            String query = "INSERT INTO pieces(color, shape, position) VALUES(?,?,?)";
-            PreparedStatement preparedStatement = con.prepareStatement(query);
+        String query = "INSERT INTO pieces(color, shape, position) VALUES(?,?,?)";
+        try (Connection con = dbConnection.getConnection();
+             PreparedStatement preparedStatement = con.prepareStatement(query)) {
             preparedStatement.setString(1, color);
             preparedStatement.setString(2, shape);
             preparedStatement.setString(3, piece.getPosition().toString());
@@ -77,10 +48,10 @@ public class PieceDAO {
     }
 
     private void deleteAll() {
-        try {
-            Connection con = getConnection();
-            String query = "DELETE FROM pieces";
-            PreparedStatement preparedStatement = con.prepareStatement(query);
+        String query = "DELETE FROM pieces";
+        try (Connection con = dbConnection.getConnection();
+             PreparedStatement preparedStatement = con.prepareStatement(query);) {
+
             preparedStatement.executeUpdate();
         } catch (SQLException throwable) {
             throwable.printStackTrace();
@@ -88,12 +59,11 @@ public class PieceDAO {
     }
 
     public Board loadPieces() {
-        try {
-            Connection con = getConnection();
-            String query = "SELECT * FROM pieces";
-            PreparedStatement preparedStatement = con.prepareStatement(query);
-            ResultSet rs = preparedStatement.executeQuery();
+        String query = "SELECT * FROM pieces";
+        try (Connection con = dbConnection.getConnection();
+             PreparedStatement preparedStatement = con.prepareStatement(query)) {
 
+            ResultSet rs = preparedStatement.executeQuery();
             List<Piece> pieces = new ArrayList<>();
 
             while (rs.next()) {
@@ -112,10 +82,9 @@ public class PieceDAO {
     }
 
     public State loadTurn(ChessGame chessGame) {
-        try {
-            Connection con = getConnection();
-            String query = "SELECT * FROM turn";
-            PreparedStatement preparedStatement = con.prepareStatement(query);
+        String query = "SELECT * FROM turn";
+        try (Connection con = dbConnection.getConnection();
+             PreparedStatement preparedStatement = con.prepareStatement(query)) {
             ResultSet rs = preparedStatement.executeQuery();
             String turn = "";
             while (rs.next()) {
@@ -133,10 +102,9 @@ public class PieceDAO {
 
     public boolean saveTurn(String status) {
         deleteTurn();
-        try {
-            Connection con = getConnection();
-            String query = "INSERT INTO turn(turn) values (?)";
-            PreparedStatement preparedStatement = con.prepareStatement(query);
+        String query = "INSERT INTO turn(turn) values (?)";
+        try (Connection con = dbConnection.getConnection();
+             PreparedStatement preparedStatement = con.prepareStatement(query)) {
             preparedStatement.setString(1, status);
             preparedStatement.executeUpdate();
             return true;
@@ -147,10 +115,9 @@ public class PieceDAO {
     }
 
     private void deleteTurn() {
-        try {
-            Connection con = getConnection();
-            String query = "DELETE FROM turn";
-            PreparedStatement preparedStatement = con.prepareStatement(query);
+        String query = "DELETE FROM turn";
+        try (Connection con = dbConnection.getConnection();
+             PreparedStatement preparedStatement = con.prepareStatement(query)){
             preparedStatement.executeUpdate();
         } catch (SQLException throwables) {
             throwables.printStackTrace();
